# contracts/internal/internal-api-v1.openapi.yaml
openapi: 3.1.0
info:
  title: SuperAgentSystem Internal API
  version: v1
  summary: Launch Spec v1 frozen internal contracts with Memory System V3 single-tenant extension
servers:
  - url: https://internal.super-agent.local
components:
  parameters:
    TenantId:
      name: X-Tenant-Id
      in: header
      required: true
      schema:
        type: string
    Scope:
      name: X-Scope
      in: header
      required: true
      schema:
        type: string
    TraceId:
      name: X-Trace-Id
      in: header
      required: true
      schema:
        type: string
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
  responses:
    ErrorResponse:
      description: Standard error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      required: [error_code, message, trace_id, retryable, details]
      properties:
        error_code:
          type: string
        message:
          type: string
        trace_id:
          type: string
        retryable:
          type: boolean
        details:
          type: object
          additionalProperties: true
...
    MemoryCandidatePreview:
      type: object
      required:
        - source_type
        - source_ref
        - artifact_tag
        - verification_status
        - side_effect_class
        - fingerprint_status
        - candidate_payload
      properties:
        source_type:
          type: string
        source_ref:
          type: string
        artifact_tag:
          type: string
        error_code:
          type: string
        verification_status:
          type: string
        side_effect_class:
          type: string
        fingerprint:
          type: string
        fingerprint_status:
          type: string
          enum: [matched, matched_or_na, mismatch, unknown]
        candidate_payload:
          type: object
          additionalProperties: true
        llm_refined_payload:
          type: object
          additionalProperties: true
    VerifyRequest:
      type: object
      required: [task_request_id, task_plan_id, verification_phase, expected_state, observed_state]
      properties:
        task_request_id:
          type: string
          format: uuid
        task_plan_id:
          type: string
          format: uuid
        task_step_id:
          type: string
          format: uuid
        verification_phase:
          type: string
          enum: [precheck, postcheck, acceptance, cleanup]
        expected_state:
          type: object
          additionalProperties: true
        observed_state:
          type: object
          additionalProperties: true
    VerifyResponse:
      type: object
      required: [verification_result_id, verdict, evidence_payload]
      properties:
        verification_result_id:
          type: string
          format: uuid
        verdict:
          type: string
          enum: [passed, failed, waived]
        failure_event_id:
          type: string
          format: uuid
        evidence_payload:
          type: object
          additionalProperties: true
        precheck_token:
          type: string
        breaker_state:
          type: string
        memory_candidate_preview:
          $ref: "#/components/schemas/MemoryCandidatePreview"
...
paths:
...
  /internal/verifier/check:
    post:
      parameters:
        - $ref: "#/components/parameters/TenantId"
        - $ref: "#/components/parameters/Scope"
        - $ref: "#/components/parameters/TraceId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyRequest"
      responses:
        "200":
          description: Verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerifyResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
...